requirements.txt
streamlit>=1.32.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
requests>=2.31.0
paho-mqtt>=1.6.1